BUG-83917. FLUME-2427. Be less verbose in isFileClosed() notification
Some versions of HDFS and implementations of FileSystem do not implement
the isFileClosed() method.

We perform a check for that method each time a new file is opened. This
patch lowers the severity of the message printed when we detect that the
method is not present from LOG.warn() to LOG.info().

We also remove the stack trace from the log message.

(Ping Wang via Mike Percy)

(cherry picked from commit 630083228364f6cd5dc18839fbd4b167cb615b38)

Change-Id: I41b2a93a32399ceae8b8b3488ed1b086252bf5af